backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
reportlab==4.2.2
gunicorn==22.0.0
psycopg[binary]==3.2.3
pytest==8.3.2